End-to-end classification problem (heart disease or not)
In this notebook I made a machine learning model to predict a patient having heart disease or not. This is my first full end-to-end project and I learned so much about it. I will continue to tune models hyperparameters to be sure the predictions will be more accurate as well as starting other projects to hone my skills further. Have a nice day!