End-to-end regression problem
In this notebook I created a machine learning model to predict the prices of bulldozers. This is dataset used for a Kaggle competition and I tried to get the model to score as good as possible on the RSMLE(Root Squared Mean Log Error). This is an demonstration of the application of a machine learning model which can be applied to a regression problem.